requirement.txt
tensorflow==2.3.0
numpy==1.18.5
efficientnet==1.1.0
streamlit==0.66.0
Pillow==7.0.0